Moving some classes shared by the receiver back to the events module
