+++ apiary-metastore-events/kafka-metastore-events/kafka-serializable-events/src/main/java/com/expediagroup/apiary/extensions/events/metastore/kafka/common/Property.java
@@ -13,24 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.apiary.extensions.events.metastore.kafka.utils;
+package com.expediagroup.apiary.extensions.events.metastore.kafka.common;
 
-public final class Preconditions {
+public interface Property {
 
-  private Preconditions() {}
+  String key();
 
-  public static String checkNotEmpty(String string, String message) {
-    if (string == null || string.trim().isEmpty()) {
-      throw new IllegalArgumentException(message);
-    }
-    return string;
-  }
+  String unPrefixedKey();
 
-  public static <T> T checkNotNull(T t, String message) {
-    if (t == null) {
-      throw new NullPointerException(message);
-    }
-    return t;
-  }
-
+  Object defaultValue();
+  
 }
